Identical testing environment created
